selectedDay now updates on user click
'Today' is marked with bold text in primary, 'selectedDay' is marked with bold text in secondary and a border. It's a bit hard to tell the difference between primary/secondary in this case, so might need to change the colors a bit. 'selectedDay' at least has a border, so that's an easy way to differentiate it from 'today'