Added planCard
"plan" section is about done from a frontend point of view -- will revisit once I set the backend up. One bug, I changed the color of the borders on the SmallDate component and now the selected border isn't showing. Probably have to add the color/borders in a css file, think it's a problem with tailwind