Updated readme, moved interfaces, tableTake2 is working
Updated readme with solutions to TS errors I was running into.
One error I found was I had the same interface (formStateVariableProps) on multiple components, but they didn't all match. So, I moved it into it's own file (src/util/interfaces.ts) and imported it everywhere. That way if I need to update it, there's only one place I need to change it.
TableTake2 is now working -- the user can enter in as many or as few ingredients as they want and they'll be formatted and added to the larger form state. Still pretty hard coded though, so might need to refactor later. Next step is to hook up the back end route to the submit button and send the inputted recipe to MongoDB.