Recipe search functionality added
Users can search for a recipe in the search bar and recipes will be filtered accordingly.

For later -- if a user types in a recipe that doesn't exist or types in gibberish, the results don't filter at all. Should update the UI somehow to reflect that the search parameters are returning nothing.